examples/with-text-embedding-models/requirements.txt
python-dotenv==1.0.0
tidb-vector>=0.0.2
langchain
openai
tiktoken